fixes #5143
